Optimize threshold decryption code
In the Baek and Zhang paper the decrypt_share step can return a valid
share or a `?` to indicate that the ciphertext was not consistent with
the private key. This `?` share was propogated forwards through the
protocol and finally filtered out in the combine_shares step.

These changes reduce the redundant verification done for decryption
shares when generating/validating and combining them. Instead this code
tracks the validation state of the ciphertext and throws exceptions if a
non validated ciphertext is used in any context other than
validate_ciphertext.