Optimize threshold decryption code
In the Baek and Zhang paper the decrypt_share step can return a valid
share or a `?` to indicate that the ciphertext was not consistent with
the private key. This `?` share was propogated forwards through the
protocol and finally filtered out in the combine_shares step.

These changes reduce the redundant verification done for decryption
shares when generating/validating and combining them. Instead this code
assumes the ciphertext supplied has been explicitly validated with
validate_ciphertext first. This requires more care when using these
functions but can save a significant amount of compute time.